refactor: extract applescript_escape into shared escape module
Move applescript_escape from terminal::common::escape to a top-level
escape module in kild-core. This eliminates the cross-domain dependency
where notify reached into terminal internals for a general-purpose
string escaping function.

Also fix pre-existing unused imports in git/health.rs exposed by
cargo fmt (moved test-only imports into #[cfg(test)] module).

Closes #327
